Pull shared functions into utils.py to be used across gpt4 and gpt
