+++ src/khoj/processor/conversation/utils.py
@@ -3,12 +3,17 @@ import logging
 from time import perf_counter
 import json
 from datetime import datetime
+import tiktoken
+
+# External packages
+from langchain.schema import ChatMessage
 
 # Internal Packages
 import queue
 from khoj.utils.helpers import merge_dicts
 
 logger = logging.getLogger(__name__)
+max_prompt_size = {"gpt-3.5-turbo": 4096, "gpt-4": 8192, "text-davinci-003": 2048}
 
 
 class ThreadedGenerator:
@@ -63,3 +68,61 @@ def message_to_log(
 
     conversation_log.extend([human_log, khoj_log])
     return conversation_log
+
+
+def generate_chatml_messages_with_context(
+    user_message, system_message, conversation_log={}, model_name="gpt-3.5-turbo", lookback_turns=2
+):
+    """Generate messages for ChatGPT with context from previous conversation"""
+    # Extract Chat History for Context
+    chat_logs = []
+    for chat in conversation_log.get("chat", []):
+        chat_notes = f'\n\n Notes:\n{chat.get("context")}' if chat.get("context") else "\n"
+        chat_logs += [chat["message"] + chat_notes]
+
+    rest_backnforths = []
+    # Extract in reverse chronological order
+    for user_msg, assistant_msg in zip(chat_logs[-2::-2], chat_logs[::-2]):
+        if len(rest_backnforths) >= 2 * lookback_turns:
+            break
+        rest_backnforths += reciprocal_conversation_to_chatml([user_msg, assistant_msg])[::-1]
+
+    # Format user and system messages to chatml format
+    system_chatml_message = [ChatMessage(content=system_message, role="system")]
+    user_chatml_message = [ChatMessage(content=user_message, role="user")]
+
+    messages = user_chatml_message + rest_backnforths + system_chatml_message
+
+    # Truncate oldest messages from conversation history until under max supported prompt size by model
+    messages = truncate_messages(messages, max_prompt_size[model_name], model_name)
+
+    # Return message in chronological order
+    return messages[::-1]
+
+
+def truncate_messages(messages, max_prompt_size, model_name):
+    """Truncate messages to fit within max prompt size supported by model"""
+    encoder = tiktoken.encoding_for_model(model_name)
+    tokens = sum([len(encoder.encode(message.content)) for message in messages])
+    while tokens > max_prompt_size and len(messages) > 1:
+        messages.pop()
+        tokens = sum([len(encoder.encode(message.content)) for message in messages])
+
+    # Truncate last message if still over max supported prompt size by model
+    if tokens > max_prompt_size:
+        last_message = "\n".join(messages[-1].content.split("\n")[:-1])
+        original_question = "\n".join(messages[-1].content.split("\n")[-1:])
+        original_question_tokens = len(encoder.encode(original_question))
+        remaining_tokens = max_prompt_size - original_question_tokens
+        truncated_message = encoder.decode(encoder.encode(last_message)[:remaining_tokens]).strip()
+        logger.debug(
+            f"Truncate last message to fit within max prompt size of {max_prompt_size} supported by {model_name} model:\n {truncated_message}"
+        )
+        messages = [ChatMessage(content=truncated_message + original_question, role=messages[-1].role)]
+
+    return messages
+
+
+def reciprocal_conversation_to_chatml(message_pair):
+    """Convert a single back and forth between user and assistant to chatml format"""
+    return [ChatMessage(content=message, role=role) for message, role in zip(message_pair, ["user", "assistant"])]
